==> 1_BP_PRS_Ensemble_prepare_phenotype_data.ipynb <==
###############################################################
# Purpose :  Prepare the Phenotype data
###############################################################

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression, Lasso, LassoCV, ElasticNet, ElasticNetCV, LogisticRegression, RidgeCV
from sklearn.metrics import r2_score, explained_variance_score, mean_squared_error, f1_score, accuracy_score
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import RandomizedSearchCV
from sklearn.model_selection import cross_val_score
from sklearn import preprocessing

import xgboost as xgb
from xgboost import XGBRegressor
import pickle

from step_1_auxiliary_functions import load_data

#load data and drop PC's from train/test split phenotype data
baseline_y_train_sbp, baseline_y_test_sbp, baseline_phenotype_data_train_sbp_with_pc, baseline_phenotype_data_test_sbp_with_pc = load_data("SBP_V1")
baseline_phenotype_data_train_sbp = baseline_phenotype_data_train_sbp_with_pc.drop(['PC_1', 'PC_2', 'PC_3', 'PC_4', 'PC_5','PC_6','PC_7','PC_8','PC_9','PC_10','PC_11'], axis=1)
baseline_phenotype_data_test_sbp = baseline_phenotype_data_test_sbp_with_pc.drop(['PC_1', 'PC_2', 'PC_3', 'PC_4', 'PC_5','PC_6','PC_7','PC_8','PC_9','PC_10','PC_11'], axis=1)

baseline_y_train_dbp, baseline_y_test_dbp, baseline_phenotype_data_train_dbp_with_pc, baseline_phenotype_data_test_dbp_with_pc = load_data("DBP_V1")
baseline_phenotype_data_train_dbp = baseline_phenotype_data_train_dbp_with_pc.drop(['PC_1', 'PC_2', 'PC_3', 'PC_4', 'PC_5','PC_6','PC_7','PC_8','PC_9','PC_10','PC_11'], axis=1)
baseline_phenotype_data_test_dbp = baseline_phenotype_data_test_dbp_with_pc.drop(['PC_1', 'PC_2', 'PC_3', 'PC_4', 'PC_5','PC_6','PC_7','PC_8','PC_9','PC_10','PC_11'], axis=1)

#export training and testing datasets
phenotype_data_export_path = '/2022_BP_ensemble/Data/TOPMed_phenotypes/baseline data/'

baseline_phenotype_data_train_sbp.to_csv(phenotype_data_export_path+"TOPMed_sbp_x_train_baseline_data.csv")
baseline_y_train_sbp.to_csv(phenotype_data_export_path+"TOPMed_sbp_y_train_baseline_data.csv")
    
baseline_phenotype_data_test_sbp.to_csv(phenotype_data_export_path+"TOPMed_sbp_x_test_baseline_data.csv")
baseline_y_test_sbp.to_csv(phenotype_data_export_path+"TOPMed_sbp_y_test_baseline_data.csv")


baseline_phenotype_data_train_dbp.to_csv(phenotype_data_export_path+"TOPMed_dbp_x_train_baseline_data.csv")
baseline_y_train_dbp.to_csv(phenotype_data_export_path+"TOPMed_dbp_y_train_baseline_data.csv")
    
baseline_phenotype_data_test_dbp.to_csv(phenotype_data_export_path+"TOPMed_dbp_x_test_baseline_data.csv")
baseline_y_test_dbp.to_csv(phenotype_data_export_path+"TOPMed_dbp_y_test_baseline_data.csv")
